Fix a warning

+++ disassembler.cpp
@@ -60,7 +60,7 @@ cs_insn* disassemble_next_branch_insn(const csh* handle, const std::vector<uint8
         // analyze disassembled instruction in @insn variable
         // NOTE: @code_ptr, @code_size & @address variables are all updated
         // to point to the next instruction after each iteration.
-        for (int i = 0; i < sizeof(branch_opcode) / sizeof(uint16_t); ++i) {
+        for (size_t i = 0; i < sizeof(branch_opcode) / sizeof(uint16_t); ++i) {
             if (insn->id == branch_opcode[i]) {
                 std::cout << "OP_STR: " << insn->op_str << std::endl;
                 return insn;
